requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-taggit==4.0.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
